% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycleways_england}
\alias{cycleways_england}
\title{Cycleways for England}
\format{
A sf object with 105184 rows and 3 variables (including geometry:
\describe{
\item{Name}{County Name}
\item{Area_Description}{county type (all the same)}
\item{geometry}{LINESTRING geometry}
}
}
\source{
\url{https://docs.ropensci.org/osmextract/}
}
\usage{
cycleways_england
}
\description{
An sf object with All cyclways for England downloaded using the example
from the {osmextract package}:
}
\keyword{datasets}
