% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_simplify.R
\name{future_simplify}
\alias{future_simplify}
\title{Parallel implementation of \code{rmapshaper::ms_simplify}}
\usage{
future_simplify(
  x,
  keep = 0.05,
  method = NULL,
  weighting = 0.7,
  keep_shapes = FALSE,
  no_repair = FALSE,
  snap = TRUE,
  explode = FALSE,
  drop_null_geometries = TRUE,
  snap_interval = NULL,
  nchunks = 1,
  .progress = FALSE
)
}
\arguments{
\item{x}{spatial object to simplify. One of:
\itemize{
\item \code{sf} or \code{sfc} polygons or lines object
}}

\item{keep}{proportion of points to retain (0-1; default 0.05)}

\item{method}{simplification method to use: \code{"vis"} for Visvalingam
algorithm, or \code{"dp"} for Douglas-Peuker algorithm. If left as
\code{NULL} (default), uses Visvalingam simplification but modifies the
area metric by underweighting the effective area of points at the vertex of
more acute angles, resulting in a smoother appearance. See this
\url{https://github.com/mbloch/mapshaper/wiki/Simplification-Tips}{link}
for more information.}

\item{weighting}{Coefficient for weighting Visvalingam simplification
(default is 0.7). Higher values produce smoother output. weighting=0 is
equivalent to unweighted Visvalingam simplification.}

\item{keep_shapes}{Prevent small polygon features from disappearing at high
simplification (default \code{FALSE})}

\item{no_repair}{disable intersection repair after simplification (default
\code{FALSE}).}

\item{snap}{Snap together vertices within a small distance threshold to fix
small coordinate misalignment in adjacent polygons. Default \code{TRUE}.}

\item{explode}{Should multipart polygons be converted to singlepart polygons?
This prevents small shapes from disappearing during simplification if
\code{keep_shapes = TRUE}. Default \code{FALSE}}

\item{drop_null_geometries}{should Features with null geometries be dropped?
Ignored for \code{Spatial*} objects, as it is always \code{TRUE}.}

\item{snap_interval}{Specify snapping distance in source units, must be a
numeric. Default \code{NULL}}

\item{nchunks}{The number of chunks to run on each core. Default is 1.
Can improve or worsen performance depending on dataset size and number of cores}

\item{.progress}{Show progress bar. Only useful when nchunks > 1. default is FALSE}
}
\value{
a simplified representation of the geometry in the same class as the
input
}
\description{
simply splits up an sf and runs the simplify algorithm on each chunk.
Uses \href{https://github.com/mbloch/mapshaper}{mapshaper} to simplify
polygons. Currently only supports sf objects.
}
