% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{gb_counties}
\alias{gb_counties}
\title{Reads the example GB counties data}
\usage{
gb_counties()
}
\value{
an sf object of all GB counties
}
\description{
Reads the example GB counties data
}
\details{
Downloaded from: "https://api.os.uk/downloads/v1/products/BoundaryLine/downloads?area=GB&format=GeoPackage&redirect"
}
\examples{
gb_counties()
}
