% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{cycleways_england}
\alias{cycleways_england}
\title{Reads the example England Cycle Network data}
\usage{
cycleways_england()
}
\value{
an sf object of the English cycle network
}
\description{
Reads the example England Cycle Network data
}
\details{
Downloaded from Open Strem Map using the {osmextract} package https://docs.ropensci.org/osmextract/
}
\examples{
cycleways_england()
}
