% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gb_counties}
\alias{gb_counties}
\title{GB Counties}
\format{
A sf object with 91 rows and 3 variables (including geometry:
\describe{
\item{Name}{County Name}
\item{Area_Description}{county type (all the same)}
\item{geometry}{MULTIPOLYGON geometry}
}
}
\source{
\url{https://osdatahub.os.uk/downloads/open/BoundaryLine}
}
\usage{
gb_counties
}
\description{
An sf object with GB counties from Ordance survey:
}
\keyword{datasets}
